changed file layout
changed file layout for ease of use since there is a lot of files now. new course viewer file added for a simpler less-A1 type of searching
